Fisrt_MRNet/requirements.txt
torch==1.8.1
paddlepaddle-gpu==2.2.1
